Исправлена ошибка в значении индекса посылки и поправлен алгоритм извлечение координаты на изображении

+++ include/sdd_protocol/connect/QSerialPortConnection.h
@@ -40,7 +40,7 @@ namespace sdd::conn {
         bool callbackIsContain(const std::function<Handle> &handler);
         State statePackageToStruct(sdd::StatePackage &pack);
 
-        int m_currentIndex;
+        int64_t m_currentIndex;
         std::shared_ptr<QSerialPort> m_serialPort;
         QMetaObject::Connection m_qtEventConnection;
         std::vector<std::function<Handle>> m_callBacks;
